Replace special-casing of weak symbol GC with generic weak pair GC
When we introduced weak pairs, it was initially rather "customized"
for the specific case of bucket chains of internal symbol tables.
This is the first step to make make it more generic, so that we can
eventually support a user-facing weak pair type.

In this change, when we mark a weak pair in the GC:
- First we convert it to a forwarding pointer (as usual)
- Then, we "recycle" the old value's car field to make it part of a
   chain of all the forwarded weak pairs.  We can get away with that
   because the pair remains allocated as 3 words, even if the
   forwarding pointer we replace it with is only 1 word.
- When GC is done, we traverse this chain and follow each forwarding
   pointer.  The car field of the weak pair in the tospace will still
   hold a pointer into fromspace, which needs to be either fixed up
   (if the target value was replaced with a forwarding pointer) or
   cleared.