Introduce a new special value for broken weak pointers
We add C_SCHEME_BROKEN_WEAK_PTR, to be used in the CAR field of a weak
pair that used to point to a now-reclaimed object.  We change this so
that later we can add a specific predicate that the user may use to
determine if the weak pair's car is no longer valid.

NOTE: It could be useful to change locatives to return this value when
the pointer is NULL, but that would be too much of a breaking change.