Add new user-facing API procedures for weak pairs to (chicken base)
This API is compatible with Chez Scheme's API for weak pairs.

Eventually, we want to make all the regular pair operators work on
weak pairs, so we do not need to add weak-c[ad]r, set-weak-c[ad]r!
etc.  This means the API is limited to three procedures:

- weak-cons to construct weak pairs
- weak-pair? to distinguish weak from regular pairs where it matters
- bwp-object? to be able to check for broken pointers in weak pairs

This commit also adds read/write invariance for #!bwp.  This is not
strictly necessary - we could decide to make broken weak pair objects
unreadable.  The advantage of this is that write won't fail randomly
just because the objects in the cars have been collected.