Tweak list operations to work for weak pairs too and document this
The list and pair operations we have typically check their argument by
comparing the header bits to the verbatim header tag.  This is extra
safe because it implicitly also checks that the type is consistent -
for instance, that it has a length of 2 and that both are block types
etc.

Instead, we add a new C accessor to grab just the type bits, ignoring
the rest, sacrificing a little bit of that safety for interoperability
between pairs and weak pairs.

Add some docs to the (chicken base) module docs, and also add a brief
mention to a few of the scheme pair operations that they accept weak
pairs.  This is not exhaustive (list?, memq/memv/member etc are not
updated) as that would be a bit silly to do.