Add new weak pair procedures to types database, and a NEWS entry
We don't want to completely overhaul the entire type system to support
weak pairs and the overloaded pair-or-weak-pair accessors which might
now return broken-weak-pointer values.  Also, the pair? predicate
could no longer be a proper predicate if we added weak pairs as
distinct types.

Instead, we accept that weak pairs are not typically used that much
and therefore it's not worth making it into a real "core" type.  With
this realisation, we can simply make the type of "(weak-cons a b)"
return "(pair (or a bwp) b)" - which is *technically* incorrect (the
bwp is replaced at GC time) but it should work because the GC might
as well happen right after (or even *during*?!) the weak-cons call.

The new bwp object, however, *is* its own distinct type, so we do need
to add it to the scrutinizer (and the lfa2 pass).

NOTE: This requires first running "make boot-chicken" and use the
resulting chicken-boot binary to compile the sources.

While we're at it, teach the non-scrutiny builtin rewriting rules in
c-platform.scm about the new procedures.