Move symbol table compaction to lookup/interning of symbols
Instead of adding an additional pass to the GC in which we scan over
the symbol tables and drop empty buckets, do this in lookup().  The
lookup() function is used by all the *intern*() functions and in
C_lookup_symbol() as well, so this is done on every symbol lookup and
interning of symbols.  The check is relatively cheap and we only
compact buckets we actually have to look at anyway.

This means we may keep around "broken" buckets for longer than
necessary, if we don't perform any lookups or interning in between
GCs.  This is probably fine - only symbol-heavy code would create lots
of broken buckets anyway, and symbol-heavy code would be doing
lookups, so we'd be doing the compaction more often as well.

This does not touch lookup_bucket() which is used to persist/unpersist
a given symbol by toggling its bucket's "is a weak pair" bit aka
specialblock bit.  We could do it there too, but I figured it makes
more sense to only do this when we have to walk the table for
string->symbol and ##sys#interned-symbol?, as those are typically
expected to be somewhat "heavy" already.

We also have to tweak compute_symbol_table_load(), as that is the
function we use in the symbolgc test to find out how many live symbols
we still have.  And more in general, it makes sense to not count
buckets with broken weak pointers as "live symbols" anyway.

While at it, drop some of the paranoid debugging checks as it would be
a bit awkward to try and shoehorn those into the lookup function().