Enable minor collection of weak pairs
In the initial weak symbol GC implementation, we added a temporary
hack in mark_nested_objects() to skip weak pairs in minor GC.

This was required because back then, we still had to scan the entire
symbol table in order to correctly drop GCed symbols.  Now that we're
not doing that anymore, we can collect weak pairs in minor GCs too.
This allows us to drop said hack.  We also can drop an extra check in
really_mark which was added to prevent adding weak pairs to the weak
pair chain in minor GC.

Drop the assert in update_weak_pairs and refactor the liveness checks
to be more comprehensive (the old checks didn't account for pairs
allocated in static memory or malloced memory - not likely, but might
be possible now they're user-facing), and to include stack checks for
the minor GC.

Change the weak pointer tests not to do major collections, and add
(slightly paranoid) checks that "permanent" symbols are kept around.
They're paranoid because permanent symbols don't actually live in
static memory - only their strings do.  So they should not result
in broken weak pointers anyway.

While working on this, I noticed that there's a small but odd
limitation - when using set-car! on a weak pair, depending on where
the pair and the car value live, we add it to the mutation stack,
which gets marked unconditionally.  Therefore, the value will be kept
alive on the next GC even if nothing else holds onto it.

To avoid this, we would need to know in C_mutate whether the slot
we're mutating is a GC-managed slot (ie it is not the first slot of a
"special block").  We don't have this information there, and making it
available would entail a massive API change.  This is not worth the
hassle, so document this minor limitation, instead.