Reset weak pair chain when switching GC mode
Whenever we build a forwarding pointer to a weak pair, we add it to
the weak pair chain.  However, when GC mode is switched, this was
done *again* for already-forwarded pointer.  The upshot of that is
that a weak pair may end up in the chain up to three times:

- First, fwded from nursery into fromspace during minor GC
- Then, fwded from fromspace to tospace during major GC
- Finally, fwded from tospace to new_fromspace during realloc GC

Since each live weak pair will be traversed and forwarded
again *anyway*, we can freely reset the weak pair chain when switching
GC modes.  This means we avoid processing the same pair more than
once, which should be faster.

This also allows us to simplify the code in update_weak_pairs, as we
don't have to chase an unknown number of forwarding pointers: there
can be only one.  We can also drop the check for the case where the
weak pointer has already been marked as broken for another copy of the
same pair in the chain.