Move line number database accessors from support.scm to expand.scm
Because support.scm is only available inside the compiler, we can't
call its procedures in the interpreter.  Since the line number
database variable itself is defined in expand.scm, it makes sense to
put the accessors there as well.

While moving, it became apparent that get-line from support.scm is
essentially the same as get-line-number from expand.scm, so use that
everywhere instead.

The other procedures are now prefixed with ##sys# because we don't
want to export them (yet?) from the user-visible chicken.syntax
module.

Rewrite display-line-number-database from printf to ##sys#print to
avoid adding a dependency on extras.scm / (chicken format) in
expand.scm.

It's still not 100% isolated - it would be better if the line number
database itself would purely be accessed through accessors, but it is
currently mutated directly in various places.