Add initial support for having line numbers available in csi
- Ensure the line number database is initialized by csi on startup
- Set up the source-info-hook for the reader so that we update the db
- Update trace emitting code to query the line number db and store it

In order to make this work properly, the trace_info_struct has to be
modified such that it can hold the location in either a raw C string
or a "cooked" Scheme string, so that we can pass the location from
Scheme (from the line number database) into the C trace buffer.

We have to keep the raw location string for emitted C code - there we
probably don't want the overhead of creating Scheme literals for every
trace-point in the code.

While at it, rename the "raw" entry to "raw_location" to clarify what
it is.  Maybe later we can rename the "cooked[12]" entries as well.

The profiler has to be updated as well to deal with the raw/cooked
location.  In case there's a cooked (Scheme string) location, we
simple emit "<eval>", so all the evaluated code ends up in the same
profiling entry.  Note that this behaviour is actually the same it was
before (as csi originally always emitted "<eval>" or "<syntax>" into
the trace buffer) - it just means that the profiler doesn't get to
benefit from the improved line number information, but it doesn't make
things worse.