Convert line number db hash table to use weak alists for form lookup
This doesn't turn it into a proper weak hash table (where the keys
would be weak), but the entries themselves are alists which are used
for further lookup of the source expression.  The symbol which is in
the car of an expression is used as the key in the hash table.

This means that the source expression may be collected.  This will
happen for instance when the symbol is redefined, like when reloading
a source file.  This means the majority of memory will be reclaimable,
and storing source locations in the interpreter will be safe.

We also clean out broken weak pointers whenever entries get added or
looked up so it doesn't grow out of proportion.  This is suboptimal,
as expressions can be added *without* traversing the alist.  This
turns O(1) entry addition into O(n) addition, or quadratic behaviour
over the number of added entries with the same car.  However, it *has*
to be done there, otherwise reloading code over and over still results
in an ever-expanding line number db.

To improve this, perhaps we can come up with some sort of "expression
hash" instead, which hashes the entire source expression, to avoid
alist usage?