Improve line number tracking in interpreter after syntax expansion
Track syntax expansion results in the line number database, just like
the compiler does.  To do this, we move update-line-number-database!
into expand.scm and prefix it with ##sys# to make it globally
accessible without having to export it from the user-visible module.
Then, we call it from eval.scm in the same way we call it from
core.scm.