Move line number tracking from csi to repl proper and update docs
This means line number tracking is also enabled in user repls, not
just the "real" csi.

When adding the read expression to the line number database, make sure
the filename isn't "#f", as it's weird to see "#f:123" as a line
number.  Instead use "<stdin>" if nothing is defined.

In the manual, drop the restriction that line-number info is only
available in the compiler.

Add a NEWS entry about the newly available line number info.