Don't override ##sys#default-read-info-hook to read with source info
Overriding this hook with fluid-let to add line number tracking has a
few problems:

- Calling (read) in csi or on other repls would cause all forms to get
  added to the line number database, even if just reading a data file.
  Even though it's not a huge problem memory-wise due to the forms
  being weakly held in the database, it could still be rather slow if
  a lot of data is read due to the scanning of broken weak pointers.
- If you use (chicken repl) in a threaded program, all (read) calls
  of the program would (unintentionally) use the line number db.

To fix this, change the code in csi as well as (load), (repl)
and (##sys#include-forms-from-file) to use ##sys#read/source-info
instead of scheme#read and fluid-letting ##sys#default-read-info-hook
to the line number tracking version.

NOTE: We could drop ##sys#default-read-info-hook entirely, but that
would break user code that uses it, like the r7rs egg, so keep it for
now.