Replace locative table with simpler "weak chain" solution
Instead of keeping track of every locative in a table, we instead use
the same approach as for tracking weak pairs: during GC, as we
encounter live locatives, build up a chain which we traverse when the
GC has completed.  We "recycle" the first slot of the locative when it
is turned into a forwarding pointer for storing the chain pointer.

Unlike weak pairs, we have to traverse both strong *and* weak
locatives, because their pointer slots need to be fixed up.  This
could be improved if we change the representation of locatives to be
object+offset instead of pointer+offset(+object), and have the
C_SPECIALBLOCK_BIT set depending on whether it is weak/strong.  This
would be a fundamental representational change, so this would be
better left for CHICKEN 6.

Signed-off-by: felix <felix@call-with-current-continuation.org>