Restore read/source-info in support.scm and export from (chicken syntax)
It turns out several eggs got broken by dropping
chicken.compiler.support#read/source-info because they are using it
directly, even though it's undocumented and supposed to be "internal".

So, restore it in support.scm as an alias for the one in
(chicken syntax) to fix those eggs, but mark it as deprecated.

Since several eggs make use of it, that settles the question of
whether it is useful to export it from (chicken syntax), so do this.
Also, since it's now an official API, make it safe to use by checking
the argument for being a port, and make the argument optional, like in
the regular (read) procedure.

Signed-off-by: felix <felix@call-with-current-continuation.org>