Fix thread id extraction in get-call-chain
This was missed when changing the trace buffer struct; all accessors
had to be adjusted by adding one to the index, except for the "raw"
slot.

Without it, the filtering of call chain entries for the current thread
will fail for anything with "frameinfo" / a filled cooked2 slot, so
the call chain will be a lot smaller than it ought to be.

Signed-off-by: felix <felix@call-with-current-continuation.org>