Don't retain weak references to finalizable objects
When an object would be garbage, but has a finalizer, it will be
artificially kept alive at the end of a GC cycle so that the finalizer
gets a chance to run, with the original event as an argument.

However, this object is effectively supposed to be garbage, and the
finalizer may put the object in a defunct state.  This means we don't
want code to be able to extract such an object through a weak
reference, because it should be considered garbage and *might* be
already cleaned up.  This could be a potential footgun, making use
after free bugs more likely.

To fix this, we make these "garbage" objects appear as garbage for the
purpose of clearing weak pointers.

The implementation is relatively trivial - we simply remember the heap
top pointer before we revive garbage objects that have finalizers, and
when dereferencing the pointer to the new object's location, where we
check that the object is in the target heap, we also check that it's
not in the portion of the heap where we've put the finalizable
objects and anything they reference.

Signed-off-by: felix <felix@call-with-current-continuation.org>