Change official name of read/source-info to avoid conflict
Somehow, when using an older compiler to bootstrap, read/source-info
from support.scm was referencing itself, even though it refers to
chicken.syntax#read-with-source-info explicitly.  This caused the
Salmonella test servers to hang.

At some point, we'll have to find out why this happened, but for now
let's take the easy way out and rename the officially exported name to
the full name "read-with-source-info".

Signed-off-by: felix <felix@call-with-current-continuation.org>