Allow overriding general installation prefix in "chicken-install"
When determining the destination location for extensions, programs, include files
and data files, respect CHICKEN_INSTALL_PREFIX (CHICKEN_INSTALL_REPOSITORY
still takes precedence for extensions). This variable was actually still
existing and used by "chicken-home" and in some places in the "chicken-install"
program. This patch removes the former use and makes this variable exclusively
a utility for "chicken-install" to override the CHICKEN installation prefix and
applies it consistently.

Signed-off-by: Peter Bex <peter@more-magic.net>