Detect redefinitions of defining forms correctly (#1132)
The scanning for local definitions in ##sys#canonicalize-body used
what I think is incorrect logic to detect whether references to
local "define-*" forms need to be expanded.

The two problems where: ##sys#macro-environment was not consulted,
so the global (default) definition would never be found to be compared
with the stored meaning in "define-definition", etc., resulting in
the fallback mode of merely testing for eq? to be used in all cases.

Second, after looking up the entry in the syntactic environment, the
value could result in a reference to another definition, so the
lookup operation needs to be repeated.

I have added test cases, as given in #1132 and removed an existing
test that seems to be wrong.

Signed-off-by: Peter Bex <peter@more-magic.net>