chicken-install: Reset egg cache when status file does not exist
If for whatever reason the status file doesn't exist, reset the cache
of the egg to prevent the object files in there from being reused.

This might happen when a CHICKEN older than 2f6a7221 caches an egg,
and we try to install that egg with a CHICKEN newer than 2f6a7221.
The older CHICKEN had installed a `STATUS' file, and the newer CHICKEN
will look for a `_STATUS' file, which won't exist.  If chicken-install
ignores the existence of the status file and reuses object files, we
might mix incompatible binaries produced by different versions of
CHICKEN and C compilers, leading to errors like

  [panic] invalid encoded literal format - execution terminated

Signed-off-by: felix <felix@call-with-current-continuation.org>