Retain current identifier status as syntax or value binding
Currently, toplevel value- and macro-bindings for an identifier are distinctly
stored in separate places, resulting in the effect that a macro definition
will shadow a value-binding (see also #1166).

One way to address this would be to remove syntax-bindings when a toplevel
identifier is "define"d and vice versa, but this will require a lot of
searching and re-consing of (possibly large) environment a-lists.

The approach chosen here is to store a global property on the symbol
that names the identifier which specifies whether a value-binding
should override any existing syntax binding (and the other way around).

Some attempt is made to properly restore the "override" status when
processing modules.

Patch updated to address definition-binding lookup loop in bodies
and ensure toplevel identifiers are correctly checked for the override
property. Also clear override-status for all imports.

Signed-off-by: Peter Bex <peter@more-magic.net>