Fix compiler warnings about zero-sized memsets
The "gethostaddr" and "fresh-addr" foreign lambdas were both declared
with "scheme-pointer" for the first "saddr" argument's type.  The
callers ("bind-socket" and "tcp-connect") both always allocate the
requisite space for the address using make-string, so this cannot
possibly be null.  This means we can use "nonnull-scheme-pointer" as
the argument's type, instead.

By not allowing the pointer to be NULL, the emitted C code uses
C_data_pointer as opposed to C_data_pointer_or_null before.  The
former doesn't have the condition which may return NULL like the
latter does, which the C compiler apparently can see, which results in
it stating that the memory size is zero (it would've been more correct
if it said that it *can be* zero).

Signed-off-by: felix <felix@call-with-current-continuation.org>