Disable test-create-temporary-file on Windows
This test fails due to the fact that "unsetenv" doesn't work like it
does on UNIX - you can only clear out an environment variable but you
can't fully unset it, and this test relies on being able to completely
erase variables from the environment.

While at it, fix the now commented-out test in the batch file to be
correct - the invocation was using UNIX style variable expansion and
didn't check its error status.  Presumably the incorrect variable
expansion would fail with a message but not stop the test due to the
missing errorlevel check.

Signed-off-by: felix <felix@call-with-current-continuation.org>