Add build-directory to rpath on MacOS
When running the checks without installing the overrides for DYLD_LIBRARY_PATH
are not seen due to changes in the way certain environment variables are
passed to subprocesses which will result in linking the wrong libchicken.dylib
at runtime. This patch adds the build directory to the rpath of the
"chicken" binary to make the checks run and changes the post-install
command for patching the rpath to the final installation directory to
add the build-path to the change list. install_name_tool accepts any
number of "-change" pairs and will silently ignore those that don't
apply.

Thanks to Christian Himpe for helping to investigate the problem and
test the fix.

Signed-off-by: Peter Bex <peter@more-magic.net>