Fix the return types of a couple of (chicken file) procedures
According to the docs, `delete-file*` is supposed to return #t on success.
However, it has `delete-file` in tail position, which in turn has `unless` in
tail position.  I think this was an oversight and `delete-file` should have
its argument in tail position instead.

The types.db incorrectly declare `delete-directory` as returning a string.

Signed-off-by: felix <felix@call-with-current-continuation.org>