Clean up process-wait return value handling; use epid for drop-child
There are three cases: either there's an error (epid = -1), there's no
child to wait for (epid = 0, if nohang was #t), or we successfully
awaited a child (epid > 0).  The original code was a bit messy and
also didn't handle the nohang situation correctly because it would
set the exit status and returned normally in the process struct if
it was passed in.  By putting the conditionals into three separate
branches, it is much clearer what the situations are.

Also, there was a bug when "proc" would be #f (wait for the next
child, whichever it is).  In that case, we would use the fallback
for get-pid (i.e, -1) to look up in the list of children, which means
we wouldn't fill in the correct struct.  Instead, use "epid", the
returned value which represents the process we just reaped.  This
would also be a memory leak because that child would never be dropped
from the list because the (unless proc ...) would only ever be called
when proc is #f, and therefore would always call (drop-child -1).

Finally, fix types.db return value types: the second
value (representing whether the child returned normally) is a boolean,
not a fixnum.  The third value is a fixnum, but may also be false when
nohang is supplied and there's nothing to wait for.