Clean up process-fork return value handling and fix ##sys#call-with-cthulhu call
Similar to 8de3923, more explicitly handle the three "pid" return
value cases: error, child or parent.

Also, wrap the call to the thunk in a lambda which is passed to
the ##sys#call-with-cthulhu procedure instead of calling it and
passing the result to ##sys#call-with-cthulhu (which somehow did not
cause an error... probably because the exit call happened inside that
thunk already) so that the continuation for the child is reset correctly.