Fix bug in process(*) auto-wait on closing of all pipes; simplify
The code in make-on-close that would call process-wait-impl after all
pipes of a subprocess are closed was basically a duplicate of what
happens in process-wait, except it had a bug - it would set the
process' exit-status field to the same value as the returned-normally?
field.

We could fix this by fixing the code here, but it's simpler to just
call chicken.process#process-wait instead.  While at it, pass the
process object to make-on-close instead of the pid, to avoid an extra
unnecessary lookup of the child.

The (input-port) and (output-port) helpers accepted a pid argument
which they didn't use; drop it.