Remove bogus safety measure
The measure used to read (unless (eq? val n) (bomb ...)), which meant
that we should not allow captured values to be rest ops.

However, with 32067934f we now do allow rest ops to refer to captured
variables.  That commit also changed the safety check, turning it into
a tautology - the val will always either be direct, or a reference to
a closure var.  So it makes no sense to even keep the check.