Add support for sharing closures to the compiler
Normally, CHICKEN uses strictly flat closures, which means that each
and every closed-over variable is wrapped in the closure object
together with the translated lambda's C function.  Unfortunately, this
can lead to a whole lot of copying if the lambda calls a lot of CPS
functions.  Each CPS call requires breaking up the lambda into
multiple different lambdas, each of which get their own C function and
closure object.

This means that a big Scheme closure which gets split up into multiple
low-level CPS closures like this will keep copying the same variables
over and over, wrapping them into a new closure, calling the CPS
function, which then "returns" to that closure, which wraps another
closure, calls another CPS function etc etc.  This can result in quite
pathological behaviour, like reported in #1852.  The C compiler also
has trouble dealing with this amount of generated code.

One could use "linked" closures which simply point to their containing
function, but that can be somewhat slow in lookup and will extend
variables' lifetime for longer than necessary, which has its own
problems.  Instead, we now create a big "container" closure when
calling the initial function, with some slots initialized as
undefined.  This container closure is then passed around to other
translated lambdas ("users") which can access the variables from that
container.  When the function which defines the variable runs, it
adds its variable to the closure at the spot reserved for it.
Subsequent "users" can then read it from the container as if they
closed over it.

This means there is no copying going on - the same container is always
passed to the next closure.  The container is the only closed-over
variable inside a sharing "user" closure.  A chain of shared closures
always starts with a "container" closure, which closes over any number
of variables.  Each subsequent user to which the container is passed
must strictly extend the set of closed-over variables.  If any
variable is dropped in a lambda, we disallow that lambda from being a
user.  Instead, it's either a regular closure or a container which
starts a new sharing chain.

We also have to take care that sharing users and containers are never
invoked more than once with the same closure object.  This would lead
to variables being clobbered.  Effectively, this means containers and
users are only created for non-escaping procedures which don't have
any mutual (direct or indirect) recursion going on.

Also, we disallow containers or users from defining more than one
sub-lambda, because that would invalidate the requirement that
closures must always be grown.  Strictly speaking, it may be possible
for a shared closure to be re-used in multiple users, but it's hard to
reason about the situations in which this would be safe and would
complicate the code too much.

Note that if a continuation is reified with call/cc, the argument to
call/cc is considered an escaping function, so it will never be marked
as a user or container.  This means it should be impossible to observe
these container slots changing around by capturing and reinvoking a
continuation (as per letrec).