Naming change in closure "transform" procedure
Change the name of the second argument in the "transform" procedure
from "here" to "crefvar".

The name "here" was a bit misleading, as "here" meant the lambda id in
both "gather" and "walk-global" in "prepare-for-code-generation",
while it referred to the closure variable ("cvar") in "transform".
This inconsistency was a bit confusing when initially trying to
understand the code.

Besides being a bit of a nitpicky improvement, this will pave the way
for the next commit, which will differentiate between the actual
closure variable (i.e., "cvar") and the variable used to access
variables inside the closure (i.e., "crefvar").

But for the current commit, the value of "crefvar" is always the same
as "cvar" which is generated when converting the lambda to a closure
literal.