Undo boxing of variables that don't escape shared closures
When a variable is created in a sharing closure container or user and
is *only* used by sharing users further down the line, and *not* any
regular lambdas (or new containers), it is not necessary to box these
variables because the container itself already acts as a box.  In such
cases, we undo the 'boxed property of the variables.