Replace quadratic lset-operations in closure generation with linear algorithms
The lset-difference/eq? and delete-duplicates calls in the closure
generation are O(n*m) and O(n^2), respectively.  When the closures are
deeply nested these might be called O(n) times, resulting in cubic
complexity of the entire closure generation step.

Implement hash-table-assisted versions of these procedures specific
for symbols so that we don't run into this problem.