Improve (chicken plist) types
On Fri, Jul 25, 2025 at 12:11:36PM +0200, Peter Bex via Chicken-hackers wrote:
> Hi there,
>
> Attached is a patch that fixes some inconsistencies between
> the definition of the procedures and types.db entries for the
> procedures defined in (chicken plist).  The commit speaks for
> itself, really.

Here's an improved version of the patch where I allow for the non-list
version of the "properties" argument to get-properties, and also
improved the third result value from "list" to "(list-of pair)".

Cheers,
Peter

From 8047fa2922e0c8b3a8f574ee50e8158a5178b700 Mon Sep 17 00:00:00 2001
From: Peter Bex <peter@more-magic.net>
Date: Fri, 25 Jul 2025 11:58:49 +0200
Subject: [PATCH] Improve (chicken plist) function type handling and
 specializations

It seems that there was an assumption that properties ("keys") are
always symbols.  However, this was not checked by either get or put!,
and the docs don't indicate that this is the case, either.

In fact, put! will allow any value type for the property, and get will
correctly return its associated value.

The only place where we checked for symbol-ness directly is in
get-properties, and only to streamline the case where the "props"
argument is not a list.  Change this to check for pair?-ness.
And while at it, we can also drop the ##sys#check-list, because props
is now guaranteed to always be a pair.

More importantly, the specializations for get and put! would only be
applied when the "key" argument is known to be a symbol, which is both
incorrect and inconvenient (sometimes a procedure will put a property
handed to it and it might not know if it's a symbol).  Change this to
allow any type.

Finally, the specialization for "get" only applied when the optional
argument was given, which is also often not the case, which means many
optimization opportunities would be missed.  Add a specialization for
the two-argument invocation.

Signed-off-by: felix <felix@call-with-current-continuation.org>