Allow safe reuse of way more closures
The original problem was where growing nested closures would be copied
over and over and over, so a lot of care needed to be taken in order
to know if the closure could be shared or not to avoid problems where
the mutations could be observed.

But closures for where there is no mutation at all it's much simpler
because they're immutable.  This happens in cases where the nested
closures simply happen to close over exactly the same variables as
their containing closure.  This means we can re-use such closures
where one closure creates multiple other closures, and where the
closures escape.